Fixed client code on server, fixed deprecated bus specifier
