+++ src/main/java/com/arrl/radiocraft/client/events/InventoryRenderEvents.java
@@ -9,7 +9,7 @@ import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.common.EventBusSubscriber;
 import net.neoforged.neoforge.client.event.ContainerScreenEvent;
 
-@EventBusSubscriber(modid = Radiocraft.MOD_ID, bus = EventBusSubscriber.Bus.GAME, value = Dist.CLIENT)
+@EventBusSubscriber(modid = Radiocraft.MOD_ID, value = Dist.CLIENT)
 public class InventoryRenderEvents {
 
     @SubscribeEvent
